Make sitemon and sendEmail into a class
