+++ main.cpp
@@ -1,0 +1,12 @@
+// ConsoleApplication2.cpp: 定义控制台应用程序的入口点。
+//
+
+#include"stdafx.h"
+#include"Sitemon.h"
+
+int main()
+{
+	Sitemon obj("www.baidu.com");
+	obj.monitor();
+	return 0;
+}
